backend/requirements.txt
# Core dependencies
express==4.18.2
cors==2.8.5
multer==1.4.5-lts.1
pdf-parse==1.1.1
docxtemplater==3.31.2
axios==1.6.0

# NLP and ML libraries (for future Python integration)
python-docx==1.2.0
PyPDF2==3.0.1
nltk==3.9.2
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.3.4
spacy==3.8.7
transformers==4.34.0
torch==2.1.0

# Development dependencies
nodemon==3.0.1